feat(front-end-web): Added Initial Loader and Loading Page (#14)
* feat(front-end-web): Added Loading Page

* fix(SelectionPage): Temporarily disable the list until types are fixed

* feat(front-end-web): Added Initial Page Loader

The initial page loader waits for cookies to register, then if there is no cookie sends the user to login. If there is, it added the user token and attempts to get the user's calendars.

* fix(MainPage): Getting Calendars No Longer Clears Auth

Fixed an error where getting calendars was clearing the auth token. Note that this does occur if it cannot fetch calendars, so it prompts the user to re-auth.

* fix(MainPage): HandleLogout now Properly Clears the Cookie