fix(auth): Authorization now gets the refresh token properly. (#20)
* fix(MainPage): Updated login/logout/load handlers to reflect oauth steps

* fix(front-end-web): Added Theme Provider

* feat(auth-url): get_auth_url now returns to localhost if that is the source

* fix(auth): Auth Flow Now Gets Refresh Token!

Partial fix to #18. Can get a refresh token if the user has not yet authorized, but sometimes a second call is made on a retry, which invalidates the token. Need to remove the retry functionality.

* fix(auth): No Double Token Calls

Closes #18. Gets the refresh token for new users every time, and loads the token from the cookie if the token is stored there.

Note: Does not get the token for users who have already authorized; that's an issue with Google. This will need to be fixed by storing tokens in the database in the future.

* fix(Main-Page): Corrected double import of "Styles"