fix(front-end-web): Streamlined Changing State
Added a handleUpdateState function that accepts just the parts of the state that are changing, so the whole state does not need to be repeated. In addition, saves parts of the state to the cookie so in the future the program can restore the user state if the page refreshes or is closed.