feat(front-end-web): Now restores state from cookie if the page reloads
The website now restores select parts of its state when the page reloads. This means that the currentUser, notification, and stage stay the same. As a result, functions like handleLogin and handleAuth now refresh the page to clean up the URL.

This also significantly improves loading performance, because it advances to the next stage of loading as soon as the previous is done, as opposed to waiting for everything to load every time.

Note that all of the functions to handle login and Google Calendar authorization have been updated to reflect this new process flow.