BREAKING CHANGE: OAuth Tokens Stored in Database, and Restore State from Cookie (#32)
* feat(front-end-web)!: Added Notification System

BREAKING CHANGE: the State.Notification interface now requires a "type".

* feat!: No longer requires cookie to store OAuth code

BREAKING CHANGE: Authorization URL has changed from `auth.html` to just `auth` to allow for processing on a single page.

Progress towards fixing issue #29

* feat!: Google Calendar OAuth Tokens Stored Directly In Database

OAuth tokens are now entirely stored in and accessed from the database. No front end cookies are used any more! This is a step forward for #29.

BREAKING CHANGE: Front End Cookie does not store OAuth tokens in cookies anymore. Therefore this breaks the function to get the calendar list.

Note: The database now supports multi-account authorization.

* fix(front-end-web): Streamlined Changing State

Added a handleUpdateState function that accepts just the parts of the state that are changing, so the whole state does not need to be repeated. In addition, saves parts of the state to the cookie so in the future the program can restore the user state if the page refreshes or is closed.

* chore: Removed extra console logging

Removed all general status console logging to clean up the code. Now only logs errors to the console.

* docs: Updated Documentation and Todo Tree Tags

* BREAKING CHANGE: Updated State Datatypes to Reflect Database Use

Datatype now properly reflects OAuth keys being stored in the database. Progress towards completing #29.

* fix(selection-page): Now Displays Calendar List Based on New Datatype

Supports multiple user accounts and the OAuth being stored in the database, but does not yet have granular privacy control. Progress on issue #29.

* feat(front-end-web): Now restores state from cookie if the page reloads

The website now restores select parts of its state when the page reloads. This means that the currentUser, notification, and stage stay the same. As a result, functions like handleLogin and handleAuth now refresh the page to clean up the URL.

This also significantly improves loading performance, because it advances to the next stage of loading as soon as the previous is done, as opposed to waiting for everything to load every time.

Note that all of the functions to handle login and Google Calendar authorization have been updated to reflect this new process flow.

* feat(front-end-web): Gets User Info from the Database on Page Load

Closes #22. Progress towards #29.

* feat: Can Now Get Calendar List Again

Can get the calendar list now that the database supports multiple accounts, and the OAuth tokens are stored in the database.

This is the last piece to close #29.