fix(front-end-web): Updated Types and Handlers to use String Types
While the use of an enum was originally preferred to ensure data consistency, problems arose with compilation and compatibility with the database. Instead everything has been updated to use the `keyof typeof` method to restrict inputs to the string values of the PrivacyTypes enum, without the restrictions presented by the enum.

Working on #31.